pyproject.toml
[tool.poetry]
name = "moviedb"
version = "0.1.0"
description = ""
authors = ["Krzysztof Świątek"]

[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.3"
gino = "^0.8.4"
starlette = "^0.12.13"
ariadne = "^0.7.0"
psycopg2-binary = "^2.8"
python-dotenv = "^0.10.3"
uvicorn = "^0.10.8"
itsdangerous = "^1.1"
aiofiles = "^0.4.0"
sqlalchemy_utils = "^0.35.0"
httpx = "^0.8.0"
awscli = "^1.16"
aiobotocore = "^0.11.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.9"
isort = "^4.3"
flake8 = "^3.7"
pytest-sugar = "^0.9.2"
pytest-cov = "^2.8.1"
pytest-env = "^0.6.2"
pytest-asyncio = "^0.10.0"
black = "^19.10b0"
faker = "^4.0.0"

[tool.black]
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
/(\.git/
  |\.eggs
  |\.hg
  |__pycache__
  |\.cache
  |\.ipynb_checkpoints
  |\.mypy_cache
  |\.pytest_cache
  |\.tox
  |\.venv
  |node_modules
  |_build
  |buck-out
  |build
  |dist
  |media
  |infrastructure
  |templates
  |locale
  |docs
)/
'''

[tool.isort]
# Vertical Hanging Indent
multi_line_output = 3
include_trailing_comma = true

line_length = 88
known_first_party = "marketplace"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

